Move Google Vertex AI task settings tests to correct package
